+++ pkg/tsdb/influxdb/models.go
@@ -8,6 +8,7 @@ type Query struct {
 	GroupBy      []*QueryPart
 	Selects      []*Select
 	RawQuery     string
+	UseRawQuery  bool
 	Alias        string
 
 	Interval string