+++ pkg/tsdb/influxdb/query.go
@@ -2,7 +2,6 @@ package influxdb
 
 import (
 	"fmt"
-	"strconv"
 	"strings"
 
 	"regexp"
@@ -11,31 +10,30 @@ import (
 )
 
 var (
-	regexpOperatorPattern *regexp.Regexp = regexp.MustCompile(`^\/.*\/$`)
+	regexpOperatorPattern    *regexp.Regexp = regexp.MustCompile(`^\/.*\/$`)
+	regexpMeasurementPattern *regexp.Regexp = regexp.MustCompile(`^\/.*\/$`)
 )
 
-type QueryBuilder struct{}
-
-func (qb *QueryBuilder) Build(query *Query, queryContext *tsdb.QueryContext) (string, error) {
-	if query.RawQuery != "" {
+func (query *Query) Build(queryContext *tsdb.QueryContext) (string, error) {
+	if query.UseRawQuery && query.RawQuery != "" {
 		q := query.RawQuery
 
-		q = strings.Replace(q, "$timeFilter", qb.renderTimeFilter(query, queryContext), 1)
+		q = strings.Replace(q, "$timeFilter", query.renderTimeFilter(queryContext), 1)
 		q = strings.Replace(q, "$interval", tsdb.CalculateInterval(queryContext.TimeRange), 1)
 
 		return q, nil
 	}
 
-	res := qb.renderSelectors(query, queryContext)
-	res += qb.renderMeasurement(query)
-	res += qb.renderWhereClause(query)
-	res += qb.renderTimeFilter(query, queryContext)
-	res += qb.renderGroupBy(query, queryContext)
+	res := query.renderSelectors(queryContext)
+	res += query.renderMeasurement()
+	res += query.renderWhereClause()
+	res += query.renderTimeFilter(queryContext)
+	res += query.renderGroupBy(queryContext)
 
 	return res, nil
 }
 
-func (qb *QueryBuilder) renderTags(query *Query) []string {
+func (query *Query) renderTags() []string {
 	var res []string
 	for i, tag := range query.Tags {
 		str := ""
@@ -59,13 +57,12 @@ func (qb *QueryBuilder) renderTags(query *Query) []string {
 		}
 
 		textValue := ""
-		numericValue, err := strconv.ParseFloat(tag.Value, 64)
 
 		// quote value unless regex or number
 		if tag.Operator == "=~" || tag.Operator == "!~" {
 			textValue = tag.Value
-		} else if err == nil {
-			textValue = fmt.Sprintf("%v", numericValue)
+		} else if tag.Operator == "<" || tag.Operator == ">" {
+			textValue = tag.Value
 		} else {
 			textValue = fmt.Sprintf("'%s'", tag.Value)
 		}
@@ -76,7 +73,7 @@ func (qb *QueryBuilder) renderTags(query *Query) []string {
 	return res
 }
 
-func (qb *QueryBuilder) renderTimeFilter(query *Query, queryContext *tsdb.QueryContext) string {
+func (query *Query) renderTimeFilter(queryContext *tsdb.QueryContext) string {
 	from := "now() - " + queryContext.TimeRange.From
 	to := ""
 
@@ -87,7 +84,7 @@ func (qb *QueryBuilder) renderTimeFilter(query *Query, queryContext *tsdb.QueryC
 	return fmt.Sprintf("time > %s%s", from, to)
 }
 
-func (qb *QueryBuilder) renderSelectors(query *Query, queryContext *tsdb.QueryContext) string {
+func (query *Query) renderSelectors(queryContext *tsdb.QueryContext) string {
 	res := "SELECT "
 
 	var selectors []string
@@ -103,19 +100,26 @@ func (qb *QueryBuilder) renderSelectors(query *Query, queryContext *tsdb.QueryCo
 	return res + strings.Join(selectors, ", ")
 }
 
-func (qb *QueryBuilder) renderMeasurement(query *Query) string {
+func (query *Query) renderMeasurement() string {
 	policy := ""
 	if query.Policy == "" || query.Policy == "default" {
 		policy = ""
 	} else {
 		policy = `"` + query.Policy + `".`
 	}
-	return fmt.Sprintf(` FROM %s"%s"`, policy, query.Measurement)
+
+	measurement := query.Measurement
+
+	if !regexpMeasurementPattern.Match([]byte(measurement)) {
+		measurement = fmt.Sprintf(`"%s"`, measurement)
+	}
+
+	return fmt.Sprintf(` FROM %s%s`, policy, measurement)
 }
 
-func (qb *QueryBuilder) renderWhereClause(query *Query) string {
+func (query *Query) renderWhereClause() string {
 	res := " WHERE "
-	conditions := qb.renderTags(query)
+	conditions := query.renderTags()
 	res += strings.Join(conditions, " ")
 	if len(conditions) > 0 {
 		res += " AND "
@@ -124,7 +128,7 @@ func (qb *QueryBuilder) renderWhereClause(query *Query) string {
 	return res
 }
 
-func (qb *QueryBuilder) renderGroupBy(query *Query, queryContext *tsdb.QueryContext) string {
+func (query *Query) renderGroupBy(queryContext *tsdb.QueryContext) string {
 	groupBy := ""
 	for i, group := range query.GroupBy {
 		if i == 0 {